+++ KeenSense/Sources/KeywordDetector/DetectedData.swift
@@ -1,8 +1,8 @@
 //
-//  PermissionAgentDelegate.swift
-//  NuguInterface
+//  DetectedData.swift
+//  KeenSense
 //
-//  Created by yonghoonKwon on 2019/11/12.
+//  Created by DCs-OfficeMBP on 19/06/2019.
 //  Copyright (c) 2019 SK Telecom Co., Ltd. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,17 +20,7 @@
 
 import Foundation
 
-/// <#Description#>
-public protocol PermissionAgentDelegate: class {
-    
-    /// <#Description#>
-    /// - Parameter names: <#names description#>
-    /// - Parameter completion: <#completion description#>
-    func permissionAgentRequestPermissions(
-        categories: Set<PermissionContext.Permission.Category>,
-        completion: @escaping () -> Void
-    )
-    
-    /// <#Description#>
-    func permissionAgentRequestContext() -> PermissionContext
+struct DetectedData {
+    let data: Data
+    let padding: Int
 }