+++ NuguClientKit/Sources/NuguClient/NuguClient+Coordinator.swift
@@ -42,7 +42,6 @@ extension NuguClient {
         setupTextAgentDependency()
         setupExtensionAgentDependency()
         setupLocationAgentDependency()
-        setupPermissionAgentDependency()
         
         // Setup core
         setupAudioStreamDependency()
@@ -139,15 +138,6 @@ extension NuguClient {
         
         contextManager.add(provideContextDelegate: agent)
     }
-    
-    func setupPermissionAgentDependency() {
-        guard let agent = permissionAgent else { return }
-        
-        agent.messageSender = networkManager
-        
-        directiveSequencer.add(handleDirectiveDelegate: agent)
-        contextManager.add(provideContextDelegate: agent)
-    }
 }
 
 // MARK: - Capability-Agents (Mandatory)
@@ -178,5 +168,7 @@ extension NuguClient {
         guard let wakeUpDetector = wakeUpDetector else { return }
         
         wakeUpDetector.audioStream = sharedAudioStream
+        
+        contextManager.add(provideContextDelegate: wakeUpDetector)
     }
 }