+++ NuguClientKit/Sources/NuguClient/NuguClient.swift
@@ -102,7 +102,7 @@ public class NuguClient {
     /// <#Description#>
     public let endPointDetector: EndPointDetectable?
     /// <#Description#>
-    public let wakeUpDetector: KeyWordDetector?
+    public let wakeUpDetector: KeywordDetector?
     
     // MARK: - Capability Agents
     
@@ -121,8 +121,6 @@ public class NuguClient {
     /// <#Description#>
     public let locationAgent: LocationAgentProtocol?
     /// <#Description#>
-    public let permissionAgent: PermissionAgentProtocol?
-    /// <#Description#>
     public let systemAgent: SystemAgentProtocol
     
     private let inputControlQueue = DispatchQueue(label: "com.sktelecom.romaine.input_control_queue")
@@ -132,15 +130,14 @@ public class NuguClient {
         inputProvider: AudioProvidable?,
         sharedAudioStream: AudioStreamable?,
         endPointDetector: EndPointDetectable?,
-        wakeUpDetector: KeyWordDetector?,
+        wakeUpDetector: KeywordDetector?,
         asrAgent: ASRAgentProtocol?,
         ttsAgent: TTSAgentProtocol?,
         audioPlayerAgent: AudioPlayerAgentProtocol?,
         displayAgent: DisplayAgentProtocol?,
         textAgent: TextAgentProtocol?,
         extensionAgent: ExtensionAgentProtocol?,
-        locationAgent: LocationAgentProtocol?,
-        permissionAgent: PermissionAgentProtocol?
+        locationAgent: LocationAgentProtocol?
     ) {
         log.info("with NuguApp")
         
@@ -155,7 +152,6 @@ public class NuguClient {
         self.textAgent = textAgent
         self.extensionAgent = extensionAgent
         self.locationAgent = locationAgent
-        self.permissionAgent = permissionAgent
         self.systemAgent = SystemAgent()
         
         setupDependencies()