+++ NuguInterface/Sources/Context/ContextPayload.swift
@@ -26,11 +26,11 @@ public struct ContextPayload {
     public let supportedInterfaces: [ContextInfo]
     /// The states of the client.
     ///
-    /// "wakeupWord" The keyword currently set in the KeyWordDetector.
+    /// "wakeupWord" The keyword currently set in the KeywordDetector.
     /// "playStack" Play list managed by PlaySyncManager.
-    public let client: [ContextInfo]?
+    public let client: [ContextInfo]
     
-    public init(supportedInterfaces: [ContextInfo], client: [ContextInfo]?) {
+    public init(supportedInterfaces: [ContextInfo], client: [ContextInfo]) {
         self.supportedInterfaces = supportedInterfaces
         self.client = client
     }