+++ SampleApp/Sources/Manager/NuguCentralManager.swift
@@ -18,12 +18,11 @@
 //  limitations under the License.
 //
 
-import Foundation
+import UIKit
 
 import NuguInterface
 import NuguClientKit
-import KeenSense
-import JadeMarble
+import NuguLoginKit
 
 final class NuguCentralManager {
     static let shared = NuguCentralManager()
@@ -33,15 +32,12 @@ final class NuguCentralManager {
     private init() {
         client.focusManager.delegate = self
         client.authorizationManager.add(stateDelegate: self)
-        client.contextManager.add(provideContextDelegate: self)
         
         if let epdFile = Bundle(for: type(of: self)).url(forResource: "skt_epd_model", withExtension: "raw") {
             client.endPointDetector?.epdFile = epdFile
         }
         
         client.locationAgent?.delegate = self
-        client.permissionAgent?.delegate = self
-        
         client.systemAgent.add(systemAgentDelegate: self)
 
         NuguLocationManager.shared.startUpdatingLocation()
@@ -65,6 +61,41 @@ extension NuguCentralManager {
         client.networkManager.disconnect()
         client.accessToken = nil
         client.inputProvider?.stop()
+    }
+    
+    func tokenRefresh() {
+        guard let refreshToken = UserDefaults.Standard.refreshToken else {
+            DispatchQueue.main.async { [weak self] in
+                SoundPlayer.playSound(soundType: .localTts(type: .deviceGatewayAuthError))
+                NuguToastManager.shared.showToast(message: "누구 앱과의 연결이 해제되었습니다. 다시 연결해주세요.")
+                self?.logout()
+            }
+            return
+        }
+        
+        OAuthManager<Type1>.shared.loginSilently(by: refreshToken) { [weak self] result in
+            switch result {
+            case .success(let authorizationInfo):
+                UserDefaults.Standard.accessToken = authorizationInfo.accessToken
+                UserDefaults.Standard.refreshToken = authorizationInfo.refreshToken
+            case .failure:
+                DispatchQueue.main.async { [weak self] in
+                    SoundPlayer.playSound(soundType: .localTts(type: .deviceGatewayAuthError))
+                    NuguToastManager.shared.showToast(message: "누구 앱과의 연결이 해제되었습니다. 다시 연결해주세요.")
+                    self?.logout()
+                }
+            }
+        }
+    }
+    
+    func logout() {
+        guard let appDelegate = UIApplication.shared.delegate as? AppDelegate,
+            let rootNavigationViewController = appDelegate.window?.rootViewController as? UINavigationController else { return }
+        DispatchQueue.main.async {
+            NuguCentralManager.shared.disable()
+            UserDefaults.Standard.clear()
+            rootNavigationViewController.popToRootViewController(animated: true)
+        }
     }
 }
 
@@ -95,15 +126,7 @@ extension NuguCentralManager {
 
 // MARK: - Internal (WakeUpDetector)
 
-extension NuguCentralManager: ContextInfoDelegate {
-    func contextInfoRequestContext() -> ContextInfo? {
-        guard let keyWord = KeyWord(rawValue: UserDefaults.Standard.wakeUpWord) else {
-            return nil
-        }
-
-        return ContextInfo(contextType: .client, name: "wakeupWord", payload: keyWord.description)
-    }
-    
+extension NuguCentralManager {
     func refreshWakeUpDetector() {
         DispatchQueue.main.async { [weak self] in
             // Should check application state, because iOS audio input can not be start using in background state
@@ -140,18 +163,32 @@ extension NuguCentralManager: ContextInfoDelegate {
     
     func setWakeUpWord(rawValue wakeUpWord: Int) {
         switch wakeUpWord {
-        case KeyWord.aria.rawValue:
-            if let netFile = Bundle.main.url(forResource: "skt_trigger_am_aria", withExtension: "raw"),
-                let searchFile = Bundle.main.url(forResource: "skt_trigger_search_aria", withExtension: "raw") {
-                client.wakeUpDetector?.netFile = netFile
-                client.wakeUpDetector?.searchFile = searchFile
+        case Keyword.aria.rawValue:
+            guard
+                let netFile = Bundle.main.url(forResource: "skt_trigger_am_aria", withExtension: "raw"),
+                let searchFile = Bundle.main.url(forResource: "skt_trigger_search_aria", withExtension: "raw") else {
+                    log.debug("keywordSource is invalid")
+                    return
             }
-        case KeyWord.tinkerbell.rawValue:
-            if let netFile = Bundle.main.url(forResource: "skt_trigger_am_tinkerbell", withExtension: "raw"),
-                let searchFile = Bundle.main.url(forResource: "skt_trigger_search_tinkerbell", withExtension: "raw") {
-                client.wakeUpDetector?.netFile = netFile
-                client.wakeUpDetector?.searchFile = searchFile
+            
+            client.wakeUpDetector?.keywordSource = KeywordSource(
+                keyword: .aria,
+                netFileUrl: netFile,
+                searchFileUrl: searchFile
+            )
+        case Keyword.tinkerbell.rawValue:
+            guard
+                let netFile = Bundle.main.url(forResource: "skt_trigger_am_tinkerbell", withExtension: "raw"),
+                let searchFile = Bundle.main.url(forResource: "skt_trigger_search_tinkerbell", withExtension: "raw") else {
+                    log.debug("keywordSource is invalid")
+                    return
             }
+            
+            client.wakeUpDetector?.keywordSource = KeywordSource(
+                keyword: .tinkerbell,
+                netFileUrl: netFile,
+                searchFileUrl: searchFile
+            )
         default:
             return
         }
@@ -190,30 +227,8 @@ extension NuguCentralManager: FocusDelegate {
 // MARK: - LocationAgentDelegate
 
 extension NuguCentralManager: LocationAgentDelegate {
-    func locationAgentRequestContext() -> LocationContext {
-        return NuguLocationManager.shared.locationContext
-    }
-}
-
-// MARK: - PermissionAgentDelegate
-
-extension NuguCentralManager: PermissionAgentDelegate {
-    func permissionAgentRequestPermissions(
-        categories: Set<PermissionContext.Permission.Category>,
-        completion: @escaping () -> Void
-    ) {
-        for category in categories {
-            switch category {
-            case .location:
-                NuguLocationManager.shared.requestLocationPermission {
-                    completion()
-                }
-            }
-        }
-    }
-    
-    func permissionAgentRequestContext() -> PermissionContext {
-        return PermissionContext(permissions: [PermissionContext.Permission(category: .location, state: NuguLocationManager.shared.permissionLocationState)])
+    func locationAgentRequestLocationInfo() -> LocationInfo? {
+        return NuguLocationManager.shared.cachedLocationInfo
     }
 }